feat: refactor OverviewContent
Split CustomPieChart and extract pure functions
